Move all encoding logic to new `encoding` directory
