+++ misc/basedirs/lib.commonjs/package.json
@@ -3,7 +3,7 @@
     "./crypto/crypto.js": "./crypto/crypto-browser.js",
     "./providers/provider-ipcsocket.js": "./providers/provider-ipcsocket-browser.js",
     "./providers/ws.js": "./providers/ws-browser.js",
-    "./utils/base64.js": "./utils/base64-browser.js",
+    "./encoding/base64.js": "./encoding/base64-browser.js",
     "./utils/geturl.js": "./utils/geturl-browser.js",
     "./wordlists/wordlists.js": "./wordlists/wordlists-browser.js"
   },
