+++ src.ts/encoding/base64.ts
@@ -5,9 +5,9 @@
  *
  * @_subsection: api/utils:Base64 Encoding  [about-base64]
  */
-import { getBytes, getBytesCopy } from './data.js';
+import { getBytes, getBytesCopy } from '../utils/data.js';
 
-import type { BytesLike } from './data.js';
+import type { BytesLike } from '../utils/data.js';
 
 /**
  * Decodes the base-64 encoded `value`.
