+++ src.ts/encoding/index.ts
@@ -1,0 +1,8 @@
+export { decodeBase58, encodeBase58 } from './base58.js';
+export { decodeBase64, encodeBase64 } from './base64.js';
+export { encodeProtoTransaction, encodeProtoWorkObject } from './proto-encode.js';
+export { decodeProtoTransaction, decodeProtoWorkObject } from './proto-decode.js';
+export { toUtf8Bytes, toUtf8CodePoints, toUtf8String, Utf8ErrorFuncs } from './utf8.js';
+
+// Types
+export type { Utf8ErrorFunc, UnicodeNormalizationForm, Utf8ErrorReason } from './utf8.js';
