+++ src.ts/encoding/proto-decode.ts
@@ -1,6 +1,6 @@
 import { ProtoTransaction } from "../transaction/abstract-transaction.js";
 import { ProtoWorkObject } from "../transaction/work-object.js";
-import * as Proto from "./ProtoBuf/proto_block.js"
+import * as Proto from "./protoc/proto_block.js"
 
 /**
  * @category Utils
