+++ src.ts/encoding/protoc/proto_block.ts
