+++ src.ts/encoding/protoc/proto_common.ts
