+++ src.ts/utils/index.ts
@@ -5,9 +5,7 @@
  * @_section api/utils:Utilities [about-utils]
  */
 
-export { decodeBase58, encodeBase58 } from './base58.js';
 
-export { decodeBase64, encodeBase64 } from './base64.js';
 
 export {
     getBytes,
@@ -56,12 +54,7 @@ export {
 
 export { resolveProperties, defineProperties } from './properties.js';
 
-export { encodeProtoTransaction, encodeProtoWorkObject } from './proto-encode.js';
-export { decodeProtoTransaction, decodeProtoWorkObject } from './proto-decode.js';
-
 export { formatQuai, parseQuai, formatEther, parseEther, formatUnits, parseUnits } from './units.js';
-
-export { toUtf8Bytes, toUtf8CodePoints, toUtf8String, Utf8ErrorFuncs } from './utf8.js';
 
 export { uuidV4 } from './uuid.js';
 
@@ -117,4 +110,3 @@ export type { FixedFormat } from './fixednumber.js';
 
 export type { BigNumberish, Numeric } from './maths.js';
 
-export type { Utf8ErrorFunc, UnicodeNormalizationForm, Utf8ErrorReason } from './utf8.js';
