--- src.ts/utils/proto.ts
@@ -1,2 +1,0 @@
-export { encodeProtoTransaction, encodeProtoWorkObject } from './proto-encode.js';
-export { decodeProtoTransaction, decodeProtoWorkObject } from './proto-decode.js';
